+++ src/Services/CRM/Item/Productrow/Result/ProductrowsResult.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Vadim Soluyanov <vadimsallee@gmail.com>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.
@@ -11,22 +11,22 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Services\CRM\Item\Result;
+namespace Bitrix24\SDK\Services\CRM\Item\Productrow\Result;
 
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Result\AbstractResult;
 
-class ItemsResult extends AbstractResult
+class ProductrowsResult extends AbstractResult
 {
     /**
-     * @return ItemItemResult[]
+     * @return ProductrowItemResult[]
      * @throws BaseException
      */
-    public function getItems(): array
+    public function getProductrows(): array
     {
         $items = [];
-        foreach ($this->getCoreResponse()->getResponseData()->getResult() as $item) {
-            $items[] = new ItemItemResult($item);
+        foreach ($this->getCoreResponse()->getResponseData()->getResult()['productRows'] as $item) {
+            $items[] = new ProductrowItemResult($item);
         }
 
         return $items;