+++ src/Services/Department/Result/DepartmentsResult.php
@@ -3,30 +3,36 @@
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Vadim Soluyanov <vadimsallee@gmail.com>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.
  */
 
+
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Services\CRM\Item\Result;
+namespace Bitrix24\SDK\Services\Department\Result;
 
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Result\AbstractResult;
 
-class ItemsResult extends AbstractResult
+/**
+ * Class DepartmentsResult
+ *
+ * @package Bitrix24\SDK\Services\Department\Result
+ */
+class DepartmentsResult extends AbstractResult
 {
     /**
-     * @return ItemItemResult[]
+     * @return DepartmentItemResult[]
      * @throws BaseException
      */
-    public function getItems(): array
+    public function getDepartments(): array
     {
         $items = [];
         foreach ($this->getCoreResponse()->getResponseData()->getResult() as $item) {
-            $items[] = new ItemItemResult($item);
+            $items[] = new DepartmentItemResult($item);
         }
 
         return $items;